requirements.txt
streamlit>=1.55
pandas
openpyxl
python-dotenv
supabase
plotly
google-genai
pdfplumber
requests
schedule
